restructure file
