+++ posthog/models/team/__init__.py
@@ -1,0 +1,1 @@
+from .team import *
