+++ posthog/models/team/team.py
@@ -1,28 +1,21 @@
 import re
-from typing import TYPE_CHECKING, Any, Dict, List, Optional
+from typing import TYPE_CHECKING, Any, List, Optional
 
 import pytz
-import structlog
 from django.contrib.postgres.fields import ArrayField
 from django.core.validators import MinLengthValidator
 from django.db import models
 
-from ee.clickhouse.sql.events import EVENTS_DATA_TABLE
-from posthog.client import sync_execute
 from posthog.constants import AvailableFeature
 from posthog.helpers.dashboard_templates import create_dashboard_from_template
+from posthog.models.dashboard import Dashboard
 from posthog.models.instance_setting import get_instance_setting
-from posthog.settings import CLICKHOUSE_CLUSTER
+from posthog.models.utils import UUIDClassicModel, generate_random_token_project, sane_repr
 from posthog.settings.utils import get_list
 from posthog.utils import GenericEmails
 
-from .dashboard import Dashboard
-from .utils import UUIDClassicModel, generate_random_token_project, sane_repr
-
 if TYPE_CHECKING:
     from posthog.models.organization import OrganizationMembership
-
-TEAM_CACHE: Dict[str, "Team"] = {}
 
 TIMEZONES = [(tz, tz) for tz in pytz.common_timezones]
 
@@ -200,30 +193,3 @@ class Team(UUIDClassicModel):
         return str(self.pk)
 
     __repr__ = sane_repr("uuid", "name", "api_token")
-
-
-logger = structlog.get_logger(__name__)
-
-# Note: Session recording, dead letter queue, logs deletion will be handled by TTL
-TABLES_TO_DELETE_FROM = lambda: [
-    EVENTS_DATA_TABLE(),
-    "person",
-    "person_distinct_id",
-    "person_distinct_id2",
-    "groups",
-    "cohortpeople",
-    "person_static_cohort",
-]
-
-
-def delete_teams_data(team_ids: List[int]):
-    logger.info(
-        f"Deleting teams data from clickhouse using background mutations.",
-        team_ids=team_ids,
-        tables=TABLES_TO_DELETE_FROM(),
-    )
-    for table in TABLES_TO_DELETE_FROM():
-        sync_execute(
-            f"ALTER TABLE {table} ON CLUSTER '{CLICKHOUSE_CLUSTER}' DELETE WHERE team_id IN %(team_ids)s",
-            {"team_ids": team_ids},
-        )
